jwt auth fix for test cases.
